+++ Runtime/DecalSystem.cs
@@ -1,112 +1,155 @@
 ﻿using UnityEngine;
+using kTools.Pooling;
 
 namespace kTools.Decals
 {
+    /// <summary>
+    /// Static system for creating and managing Decals.
+    /// </summary>
     public static class DecalSystem
     {
+#region Pool
         /// <summary>
-        /// Get a Decal instance.
+        /// Create Pool for Decals using DecalData as key.
         /// </summary>
-        /// <param name="positionWS">Decal position in World space.</param>
-        /// <param name="directionWS">World space direction/normal vector to use for Decal rotation.</param>
-        /// <param name="decalData">DecalData to set.</param>
-        /// <param name="usePooling">If true the Decal will be taken from a DecalPooler instance.</param>
-        public static Decal GetDecal(Vector3 positionWS, Vector3 directionWS, ScriptableDecal decalData, bool usePooling)
+        /// <param name="decalData">DecalData to create Decals from.</param>
+        public static void CreateDecalPool(DecalData decalData)
         {
-			return GetDecal(positionWS, directionWS, Vector2.one, decalData, usePooling);
-        }
-
-        /// <summary>
-        /// Get a Decal instance.
-        /// </summary>
-        /// <param name="positionWS">Decal position in World space.</param>
-        /// <param name="directionWS">World space direction/normal vector to use for Decal rotation.</param>
-        /// <param name="scaleWS">Decal scale in World space.</param>
-        /// <param name="decalData">DecalData to set.</param>
-        /// <param name="usePooling">If true the Decal will be taken from a DecalPooler instance.</param>
-		public static Decal GetDecal(Vector3 positionWS, Vector3 directionWS, Vector2 scaleWS, ScriptableDecal decalData, bool usePooling)
-        {
-            Decal decal = GetDecalInstance(decalData, usePooling);
-            decal.SetActive(true);
-            decal.SetTransform(positionWS, directionWS, scaleWS);
-            decal.SetData(decalData);
-			return decal;
-        }
-
-        /// <summary>
-        /// Get a Decal instance.
-        /// </summary>
-        /// <param name="positionWS">Decal position in World space.</param>
-        /// <param name="rotationWS">Decal rotation in World space.</param>
-        /// <param name="decalData">DecalData to set.</param>
-        /// <param name="usePooling">If true the Decal will be taken from a DecalPooler instance.</param>
-        public static Decal GetDecal(Vector3 positionWS, Quaternion rotationWS, ScriptableDecal decalData, bool usePooling)
-        {
-			return GetDecal(positionWS, rotationWS, Vector2.one, decalData, usePooling);
-        }
-
-        /// <summary>
-        /// Get a Decal instance.
-        /// </summary>
-        /// <param name="positionWS">Decal position in World space.</param>
-        /// <param name="rotationWS">Decal rotation in World space.</param>
-        /// <param name="scaleWS">Decal scale in World space.</param>
-        /// <param name="decalData">DecalData to set.</param>
-        /// <param name="usePooling">If true the Decal will be taken from a DecalPooler instance.</param>
-		public static Decal GetDecal(Vector3 positionWS, Quaternion rotationWS, Vector2 scaleWS, ScriptableDecal decalData, bool usePooling)
-        {
-            Decal decal = GetDecalInstance(decalData, usePooling);
-            decal.SetActive(true);
-            decal.SetTransform(positionWS, rotationWS, scaleWS);
-            decal.SetData(decalData);
-			return decal;
-        }
-
-        /// <summary>
-        /// Create a new Decal directly. Does not handle Transform or Renderer setup.
-        /// </summary>
-        /// <param name="decalData">DecalData to set.</param>
-        public static Decal CreateDecalDirect(ScriptableDecal decalData)
-        {
-            GameObject obj = new GameObject();
-            obj.name = string.Format("Decal_{0}", decalData.name);
-            Transform transform = obj.transform;
-            transform.position = Vector3.zero;
-            transform.rotation = Quaternion.identity;
-            transform.localScale = Vector3.one;
-            return transform.gameObject.AddComponent<Decal>();
-        }
-
-        // -------------------------------------------------- //
-        //                  INTERNAL METHODS                  //
-        // -------------------------------------------------- //
-
-        // Get a Decal instance, either from pooling or directly
-        private static Decal GetDecalInstance(ScriptableDecal decalData, bool usePooling)
-        {
-            if(usePooling)
+            // Test for pooling enabled
+            if(!decalData.poolingEnabled)
             {
-                // Ensure there is an active DecalPooler
-                DecalPooler decalPooler = GameObject.FindObjectOfType<DecalPooler>();
-                if(decalPooler == null)
-                    CreateDecalPooler();
+                Debug.LogWarning($"Pooling is not enabled for DecalData ({decalData.name})");
+                return;
+            }
 
-                // Get a poolable Decal instance
+            // Create Pool
+            var decal = CreateDecal(decalData);
+            PoolingSystem.CreatePool(decalData, decal, decalData.instanceCount);
+
+            // Cleanup
+            DestroyGameObject(decal.gameObject);
+        }
+
+        /// <summary>
+        /// Destroy Pool of Decals.
+        /// </summary>
+        /// <param name="decalData">Key for Pool to destroy.</param>
+        public static void DestroyDecalPool(DecalData decalData)
+        {
+            // Test for pooling enabled
+            if(!decalData.poolingEnabled)
+            {
+                Debug.LogWarning($"Pooling is not enabled for DecalData ({decalData.name})");
+                return;
+            }
+
+            // Destroy Pool
+            PoolingSystem.DestroyPool<Decal>(decalData);
+        }
+
+        /// <summary>
+        /// Tests whether Pool exists for DecalData.
+        /// </summary>
+        /// <param name="decalData">Key to test for.</param>
+        /// <returns>True if Pool exists.</returns>
+        public static bool HasDecalPool(DecalData decalData)
+        {
+            // Test for pooling enabled
+            if(!decalData.poolingEnabled)
+            {
+                Debug.LogWarning($"Pooling is not enabled for DecalData ({decalData.name})");
+                return false;
+            }
+
+            // Test for matching Pool
+            return PoolingSystem.HasPool<Decal>(decalData);
+        }
+#endregion
+
+#region Decal
+        /// <summary>
+        /// Get new Decal and set Transform. If Pooling is enabled on DecalData, Decal will be taken from Pool.
+        /// </summary>
+        /// <param name="decalData">DecalData to create Decal from.</param>
+        /// <param name="position">World space position for Decal.</param>
+        /// <param name="direction">World space forward direction for Decal.</param>
+        /// <param name="scale">Local space scale for Decal.</param>
+        /// <returns></returns>
+        public static Decal GetDecal(DecalData decalData, Vector3 position, Vector3 direction, Vector3 scale)
+        {
+            var decal = GetDecal(decalData);
+            decal.SetTransform(position, direction, scale);
+            return decal;
+        }
+
+        /// <summary>
+        /// Get new Decal. If Pooling is enabled on DecalData, Decal will be taken from Pool.
+        /// </summary>
+        /// <param name="decalData">DecalData to create Decal from.</param>
+        /// <returns></returns>
+        public static Decal GetDecal(DecalData decalData)
+        {
+            // Test for pooling enabled
+            if(decalData.poolingEnabled)
+            {
+                // Create Pool
+                if(!HasDecalPool(decalData))
+                {
+                    CreateDecalPool(decalData);
+                }
+
+                // Get Decal from Pool
                 Decal decal;
-                DecalPooler.Instance.TryGetInstance(decalData, out decal);
+                PoolingSystem.TryGetInstance(decalData, out decal);
                 return decal;
             }
-            else
-                // Create single Decal
-                return CreateDecalDirect(decalData);
+            
+            // Create new Decal
+            return CreateDecal(decalData);
+        }
+        
+        /// <summary>
+        /// Remove existing Decal. If Pooling is enabled on DecalData, Decal will be returned to Pool.
+        /// </summary>
+        /// <param name="decal">Decal to remove.</param>
+        public static void RemoveDecal(Decal decal)
+        {
+            // Test for pooling enabled
+            var key = decal.decalData;
+            if(key.poolingEnabled)
+            {
+                // Try to return Decal to Pool
+                if(HasDecalPool(key))
+                {
+                    PoolingSystem.ReturnInstance<Decal>(key, decal);
+                    return;
+                }
+            }
+
+            // Destroy decal
+            DestroyGameObject(decal.gameObject);
         }
 
-        // Create a new DecalPooler instance
-        private static void CreateDecalPooler()
+        static Decal CreateDecal(DecalData decalData)
         {
-            GameObject obj = new GameObject();
-            obj.name = "DecalPooler";
-            obj.AddComponent<DecalPooler>();
+            // Create new Decal
+            var name = decalData.name;
+            var obj = new GameObject(name, typeof(Decal));
+            var decal = obj.GetComponent<Decal>();
+            decal.Init(decalData);
+            return decal;
         }
+#endregion
+
+#region GameObject
+        static void DestroyGameObject(GameObject gameObject)
+        {
+            #if UNITY_EDITOR
+            Object.DestroyImmediate(gameObject);
+            #else
+            Object.Destroy(gameObject);
+            #endif
+        }
+#endregion
     }
 }